refactor dids
