+++ src/common/version.ts
@@ -1,1 +1,1 @@
-export const VERSION = "0.24.2-alpha-2"
+export const VERSION = "0.24.2-alpha-3"
