+++ src/did/index.ts
@@ -1,0 +1,4 @@
+export * from './getters'
+export * from './transformers'
+export * from './types'
+export * from './validation'
