+++ src/did/types.ts
@@ -1,0 +1,5 @@
+export enum KeyType {
+  RSA = 'rsa',
+  Edwards = 'ed25519',
+  BLS = 'bls12-381'
+}
